+++ src/config/info.ts
@@ -1,4 +1,4 @@
 export default {
-    version: "2.5.1",
+    version: "2.6.0",
     sharkyID: "151044827738275840"
 };